Finish decoupling wayland_display and flutter_application
